requirements.txt
python-telegram-bot==20.7
telethon
sentence-transformers
faiss-cpu
python-dotenv
requests 